+++ src/web/mrseriesqa.php
@@ -47,7 +47,7 @@
 	require "includes_php.php";
 	require "includes_html.php";
 
-	$id = GetVariable("id");
+	$id = (int)GetVariable("id");
 
 	/* get the path to the QA info */
 	$sqlstring = "select a.*, b.study_num, d.uid from mr_series a left join studies b on a.study_id = b.study_id left join enrollment c on b.enrollment_id = c.enrollment_id left join subjects d on c.subject_id = d.subject_id where a.mrseries_id = $id";
