+++ src/web/templates.php
@@ -52,9 +52,9 @@
 
 	/* ----- setup variables ----- */
 	$action = GetVariable("action");
-	$templateid = GetVariable("templateid");
-	$projectid = GetVariable("projectid");
-	$ptid = GetVariable("ptid");
+	$templateid = (int)GetVariable("templateid");
+	$projectid = (int)GetVariable("projectid");
+	$ptid = (int)GetVariable("ptid");
 	$newtemplatename = GetVariable("newtemplatename");
 	$newtemplatemodality = GetVariable("newtemplatemodality");
 	$visittype = GetVariable("visittype");
