+++ src/web/upload.php
@@ -200,8 +200,8 @@ class qqFileUploader {
 	
 /* params are stored in $_GET */
 $modality = strtolower(GetVariable("modality"));
-$studyid = GetVariable("studyid");
-$seriesid = GetVariable("seriesid");
+$studyid = (int)GetVariable("studyid");
+$seriesid = (int)GetVariable("seriesid");
 
 if ($modality == "mr") {
 	$uploadpath = $GLOBALS['dicomincomingpath'] . '/';
